Added working UserPermissionRequest features.

+++ app/controllers/babies_controller.rb
@@ -10,6 +10,11 @@ class BabiesController < ApplicationController
     @baby = Baby.new(params[:baby])
     @baby.owner = user
     @baby.users << user
+    #add the users who have previously had permission requests approved
+    users_to_add = UserPermissionRequest.where(:requested_of_id => user.id, :status => UserPermissionRequest::STATUS_APPROVED)
+    users_to_add.each do |u|
+      @baby.users << u.requested_by unless (@baby.owner == u.requested_by || @baby.users.include?(u.requested_by))
+    end
     @baby.log = Log.create(:creator => user)
     @baby.save
     flash[:notice] = "Baby created!"
@@ -36,8 +41,12 @@ class BabiesController < ApplicationController
   
   def destroy
     @baby = Baby.find_by_id(params[:id])
-    if @baby.destroy 
-      flash[:notice] = "Baby record removed."
+    if @baby.owner == User.find_by_id(session[:user_id])
+      if @baby.destroy 
+        flash[:notice] = "Baby record removed."
+      end
+    else
+      flash[:notice] = "No way, Jose!  You're not that baby's daddy (or mommy!)"
     end
     redirect_to :controller => :users, :action => :home
   end
